requirements.txt
python==3.9
pysam==0.17.0
samtools==1.12
impyla==0.16.3
thriftpy2==0.4.14
thrift==0.13.0
lark-parser==0.6.7
networkx==2.6.2
numpy==1.21.2
scipy==1.7.1
scikit-learn==1.0.1
matplotlib-base==3.4.2
pandas==1.3.4
pyarrow==6.0.1
aiohttp==3.8.1
fsspec==2021.11.1
s3fs==2021.11.1
pyliftover==0.4
python-box==5.3.0
setuptools==52.0.0
sqlalchemy==1.4.27
sqlite==3.36.0
django==3.2.12
django-cors-headers==3.7.0
django-guardian==2.4.0
djangorestframework==3.12.4
openjdk==8.0.152
hadoop==3.1.2
deprecation==2.1.0
toml==0.10.2
cerberus==1.3.4
jinja2==3.0.1
snakemake-minimal==6.12.1
ijson==3.1.4
dask==2022.1.0
dask-kubernetes==2022.1.0
bump2version==1.0.1
pytest-doctestplus==0.4.0
sphinx==4.1.2
sphinx_rtd_theme==0.5.2
sphinx-copybutton==0.4.0
sphinxcontrib-httpdomain==1.7.0
coverage==5.5
flake8==3.9.2
mypy==0.910
pytest==6.2.4
pytest-cov==2.12.1
pytest-mock==3.6.1
pytest-xdist==2.3.0
pytest-django==4.4.0
pylint==2.9.1
pre-commit==2.13.0
moto==2.3.0
pandas-stubs==1.2.0.43
django-stubs==1.9.0
types-pyyaml==6.0.3
types-setuptools==57.4.7
sqlalchemy-stubs==0.4
data-science-types==0.2.23
types-toml==0.10.3
